alpha_miner/requirements.txt
openai
anthropic
faiss-cpu
sqlalchemy
pydantic
tenacity
pandas
numpy
scipy
cvxpy
pyyaml
pytest
pyarrow
